+++ packages/blockchain-link/tests/unit/fixtures/getAccountInfo-blockbook.ts
@@ -181,9 +181,9 @@ export default [
                                 vout: [
                                     {
                                         addresses: ['1J8tVQD9KZZeLhnkMRHHDawsYmwjWAnC5d'],
+                                        value: '1',
                                     },
                                 ],
-                                value: '1',
                             },
                         ],
                         page: 1,
@@ -246,8 +246,8 @@ export default [
                     data: {
                         address:
                             'xpub6CVKsQYXc9b2MiuW1HisiJKCEyB8vSqEafi6CP6Qu96YABCKScWtm1gUko1yDRSdmPjYQ8eFUbc9qrvNxBTUq2Z19aenEmCFcUcFUJL1Wpu',
-                        balance: '1',
-                        unconfirmedBalance: '-1',
+                        balance: '100',
+                        unconfirmedBalance: '-60', // output.value + fee
                         txs: 2,
                         tokens: [
                             {
@@ -256,21 +256,33 @@ export default [
                                 name: '1J8tVQD9KZZeLhnkMRHHDawsYmwjWAnC5d',
                                 transfers: 2,
                             },
+                            {
+                                type: 'XPUBAddress',
+                                path: "m/44'/0'/100'/1/1",
+                                name: '1RXiBGixLSBRAAXtZMsCx75EuFqqJnmXZ',
+                                transfers: 1,
+                            },
                         ],
                         transactions: [
                             {
                                 vin: [
                                     {
                                         addresses: ['1J8tVQD9KZZeLhnkMRHHDawsYmwjWAnC5d'],
+                                        value: '100',
                                         sequence: 4294967293, // RBF
                                     },
                                 ],
                                 vout: [
                                     {
                                         addresses: ['19SW698tGLusJZVBmGDYmHvSwn79WqJP65'],
+                                        value: '50',
+                                    },
+                                    {
+                                        addresses: ['1RXiBGixLSBRAAXtZMsCx75EuFqqJnmXZ'],
+                                        value: '40',
                                     },
                                 ],
-                                value: '1',
+                                fees: '10',
                             },
                         ],
                         page: 1,
@@ -284,18 +296,20 @@ export default [
             descriptor:
                 'xpub6CVKsQYXc9b2MiuW1HisiJKCEyB8vSqEafi6CP6Qu96YABCKScWtm1gUko1yDRSdmPjYQ8eFUbc9qrvNxBTUq2Z19aenEmCFcUcFUJL1Wpu',
             empty: false,
-            balance: '1',
-            availableBalance: '0',
+            balance: '100',
+            availableBalance: '40',
             history: {
                 total: 2,
                 transactions: [
                     {
                         type: 'sent',
-                        amount: '1',
+                        amount: '60',
+                        fee: '10',
                         rbf: true,
                         targets: [
                             {
                                 addresses: ['19SW698tGLusJZVBmGDYmHvSwn79WqJP65'],
+                                amount: '50',
                             },
                         ],
                         tokens: [],
@@ -308,6 +322,11 @@ export default [
                         path: "m/44'/0'/100'/1/0",
                         address: '1J8tVQD9KZZeLhnkMRHHDawsYmwjWAnC5d',
                         transfers: 2,
+                    },
+                    {
+                        path: "m/44'/0'/100'/1/1",
+                        address: '1RXiBGixLSBRAAXtZMsCx75EuFqqJnmXZ',
+                        transfers: 1,
                     },
                 ],
                 used: [],