+++ packages/suite/src/reducers/wallet/sendFormReducer.ts
@@ -1,8 +1,9 @@
 import produce from 'immer';
 import validator from 'validator';
+import Bignumber from 'bignumber.js';
 import { SEND } from '@wallet-actions/constants';
 import { getOutput } from '@wallet-utils/sendFormUtils';
-import { State, Output } from '@wallet-types/sendForm';
+import { State, InitialState, Output } from '@wallet-types/sendForm';
 import {
     VALIDATION_ERRORS,
     FIRST_OUTPUT_ID,
@@ -11,7 +12,7 @@ import {
 import { isAddressValid } from '@wallet-utils/validation';
 import { WalletAction } from '@wallet-types';
 
-export const initialState: State = {
+const initialState = (loaded: InitialState): State => ({
     outputs: [
         {
             // fill first output by default
@@ -22,7 +23,6 @@ export const initialState: State = {
             localCurrency: { value: DEFAULT_LOCAL_CURRENCY },
         },
     ],
-    fee: null,
     customFee: { value: null, error: null },
     isAdditionalFormVisible: false,
     networkTypeRipple: {
@@ -36,10 +36,16 @@ export const initialState: State = {
         gasLimit: { value: null, error: null },
         data: { value: null, error: null },
     },
-    networkTypeBitcoin: {},
-};
+    networkTypeBitcoin: {
+        transactionInfo: null,
+    },
+    ...loaded,
+});
 
-export default (state: State = initialState, action: WalletAction): State => {
+export default (state: State | null = null, action: WalletAction): State | null => {
+    if (action.type === SEND.INIT) return initialState(action.payload);
+    if (!state || action.type === SEND.DISPOSE) return null;
+
     return produce(state, draft => {
         switch (action.type) {
             // show additional form
@@ -102,17 +108,23 @@ export default (state: State = initialState, action: WalletAction): State => {
             }
 
             // change select "Fee"
-            case SEND.HANDLE_FEE_VALUE_CHANGE: {
-                const { fee } = action;
-                draft.fee = fee;
+            case SEND.HANDLE_FEE_VALUE_CHANGE:
+                draft.selectedFee = action.fee;
                 break;
-            }
 
             // change select "Fee"
             case SEND.HANDLE_CUSTOM_FEE_VALUE_CHANGE: {
                 const { customFee } = action;
+
                 draft.customFee.error = null;
                 draft.customFee.value = customFee;
+                draft.selectedFee.feePerUnit = customFee || '1';
+
+                if (customFee === null) return draft;
+
+                const customFeeBig = new Bignumber(customFee);
+                const maxFeeBig = new Bignumber(draft.feeInfo.maxFee);
+                const minFeeBig = new Bignumber(draft.feeInfo.minFee);
 
                 if (validator.isEmpty(customFee)) {
                     draft.customFee.error = VALIDATION_ERRORS.IS_EMPTY;
@@ -123,6 +135,12 @@ export default (state: State = initialState, action: WalletAction): State => {
                     draft.customFee.error = VALIDATION_ERRORS.NOT_NUMBER;
                     return draft;
                 }
+
+                if (customFeeBig.isGreaterThan(maxFeeBig) || customFeeBig.isLessThan(minFeeBig)) {
+                    draft.customFee.error = VALIDATION_ERRORS.NOT_IN_RANGE;
+                    return draft;
+                }
+
                 break;
             }
 
@@ -146,16 +164,27 @@ export default (state: State = initialState, action: WalletAction): State => {
                 break;
             }
 
+            // click button "Add recipient"
+            case SEND.BTC_REMOVE_RECIPIENT: {
+                const { outputId } = action;
+                const removed = draft.outputs.filter(output => output.id !== outputId);
+                draft.outputs = removed;
+                break;
+            }
+
             // click button "Clear"
             case SEND.CLEAR: {
-                return {
-                    ...initialState,
+                return initialState({
+                    feeInfo: draft.feeInfo,
+                    selectedFee:
+                        draft.feeInfo.levels.find(f => f.label === 'normal') ||
+                        draft.feeInfo.levels[0],
                     isAdditionalFormVisible: draft.isAdditionalFormVisible,
-                };
+                });
             }
 
             // change input in additional xrp form "Destination tag"
-            case SEND.HANDLE_XRP_DESTINATION_TAG_CHANGE: {
+            case SEND.XRP_HANDLE_DESTINATION_TAG_CHANGE: {
                 const { destinationTag } = action;
                 draft.networkTypeRipple.destinationTag.error = null;
                 draft.networkTypeRipple.destinationTag.value = destinationTag;
@@ -166,6 +195,10 @@ export default (state: State = initialState, action: WalletAction): State => {
                 }
                 break;
             }
+
+            case SEND.BTC_PRECOMPOSED_TX:
+                draft.networkTypeBitcoin.transactionInfo = action.payload;
+                break;
 
             // no default
         }