+++ packages/suite/src/types/wallet/index.ts
@@ -3,6 +3,7 @@ import { ReceiveActions } from '@wallet-actions/receiveActions';
 import { SignVerifyActions } from '@wallet-actions/signVerifyActions';
 
 import { SendFormActions } from '@wallet-actions/sendFormActions';
+import { SendFormCacheActions } from '@wallet-actions/sendFormCacheActions';
 import { SendFormBitcoinActions } from '@wallet-actions/sendFormSpecific/bitcoinActions';
 import { SendFormRippleActions } from '@wallet-actions/sendFormSpecific/rippleActions';
 
@@ -11,7 +12,6 @@ import { AccountActions } from '@wallet-actions/accountActions';
 import { Discovery as Discovery$ } from '@wallet-reducers/discoveryReducer';
 import { Account as Account$ } from '@wallet-reducers/accountsReducer';
 import { Fiat as Fiat$ } from '@wallet-reducers/fiatRateReducer';
-import { Fee as Fee$ } from '@wallet-reducers/feesReducer';
 
 import { FiatRateActions } from '@wallet-middlewares/coingeckoMiddleware';
 import { BlockchainActions } from '@wallet-actions/blockchainActions';
@@ -26,7 +26,6 @@ export type Network = ArrayElement<typeof NETWORKS>;
 export type ExternalNetwork = ArrayElement<typeof EXTERNAL_NETWORKS>;
 
 // this weird export is because of --isolatedModules and next.js 9
-export type Fee = Fee$;
 export type NetworkToken = NetworkToken$;
 export type Token = Token$;
 export type Account = Account$;
@@ -45,6 +44,7 @@ export type WalletAction =
     | SettingsActions
     | ReceiveActions
     | SendFormActions
+    | SendFormCacheActions
     | SendFormBitcoinActions
     | SendFormRippleActions
     | SignVerifyActions