+++ packages/suite/src/views/wallet/account/send/components/AdditionalForm/components/CustomFee/index.tsx
@@ -4,7 +4,8 @@ import { FormattedMessage } from 'react-intl';
 import { Input } from '@trezor/components';
 import sendMessages from '@wallet-views/account/messages';
 import { VALIDATION_ERRORS } from '@wallet-constants/sendForm';
-import messages from '../../index.messages';
+import { State } from '@wallet-types/sendForm';
+import messages from './index.messages';
 import { Props as ContainerProps } from '../../../../Container';
 
 const Label = styled.div`
@@ -14,15 +15,14 @@ const Label = styled.div`
 `;
 
 interface Props {
-    errors: ContainerProps['send']['customFee']['error'];
-    customFee: ContainerProps['send']['customFee']['value'];
+    errors: State['customFee']['error'];
+    customFee: State['customFee']['value'];
+    maxFee: State['feeInfo']['maxFee'];
+    minFee: State['feeInfo']['minFee'];
     sendFormActions: ContainerProps['sendFormActions'];
 }
 
-const getState = (
-    error: ContainerProps['send']['customFee']['error'],
-    customFee: ContainerProps['send']['customFee']['value'],
-) => {
+const getState = (error: Props['errors'], customFee: Props['customFee']) => {
     if (error) {
         return 'error';
     }
@@ -31,12 +31,23 @@ const getState = (
     }
 };
 
-const getErrorMessage = (error: ContainerProps['send']['customFee']['error']) => {
+const getErrorMessage = (
+    error: Props['errors'],
+    maxFee: Props['maxFee'],
+    minFee: Props['minFee'],
+) => {
     switch (error) {
         case VALIDATION_ERRORS.IS_EMPTY:
             return <FormattedMessage {...messages.TR_CUSTOM_FEE_IS_NOT_SET} />;
         case VALIDATION_ERRORS.NOT_NUMBER:
             return <FormattedMessage {...messages.TR_CUSTOM_FEE_IS_NOT_VALID} />;
+        case VALIDATION_ERRORS.NOT_IN_RANGE:
+            return (
+                <FormattedMessage
+                    {...messages.TR_CUSTOM_FEE_NOT_IN_RANGE}
+                    values={{ maxFee, minFee }}
+                />
+            );
         default:
             return null;
     }
@@ -50,7 +61,7 @@ const CustomFee = (props: Props) => (
                 <FormattedMessage {...sendMessages.TR_FEE} />
             </Label>
         }
-        bottomText={getErrorMessage(props.errors)}
+        bottomText={getErrorMessage(props.errors, props.maxFee, props.minFee)}
         value={props.customFee || ''}
         onChange={e => {
             props.sendFormActions.handleCustomFeeValueChange(e.target.value);