+++ packages/typescript-typings/types/trezor-connect/index.d.ts
@@ -276,6 +276,42 @@ declare module 'trezor-connect' {
         push?: boolean;
     }
 
+    export interface PrecomposeTransactionParams extends CommonParams {
+        outputs: Output[];
+        account: {
+            path: string;
+            addresses: AccountAddresses;
+            utxo: AccountUtxo[];
+        }
+        feeLevels: {
+            feePerUnit: string,
+        }[];
+        coin: string;
+    }
+
+    export type PrecomposedTransaction = {
+        type: 'error',
+        error: string,
+    } | {
+        type: 'nonfinal',
+        max: string,
+        totalSpent: string, // all the outputs, no fee, no change
+        fee: string,
+        feePerByte: string,
+        bytes: number,
+    } | {
+        type: 'final',
+        max: string,
+        totalSpent: string, // all the outputs, no fee, no change
+        fee: string,
+        feePerByte: string,
+        bytes: number,
+        transaction: {
+            inputs: TransactionInput[],
+            outputs: TransactionOutput[],
+        },
+    }
+
     export interface Transaction {
         signatures: string[]; // signer signatures
         serializedTx: string; // serialized transaction
@@ -379,7 +415,7 @@ declare module 'trezor-connect' {
     export interface SignedTransaction {
         signatures: string[];
         serializedTx: string;
-        txId?: string;
+        txid?: string;
     }
 
     export interface Settings {
@@ -437,8 +473,16 @@ declare module 'trezor-connect' {
         type: 'opreturn';
         dataHex: string;
     }
+    export interface NoAddressOutput {
+        type: 'noaddress';
+        amount: string;
+    }
 
-    export type Output = RegularOutput | InternalOutput | SendMaxOutput | OpReturnOutput;
+    export interface NoAddressSendMaxOutput {
+        type: 'send-max-noaddress';
+    }
+
+    export type Output = RegularOutput | InternalOutput | SendMaxOutput | OpReturnOutput | NoAddressOutput | NoAddressSendMaxOutput;
 
     export interface BinOutput {
         amount: number;
@@ -774,12 +818,41 @@ declare module 'trezor-connect' {
         rskip60?: number; // eth
     }
     
+    interface BlockchainEstimateFeeParams {
+        coin: string;
+        request?: {
+            blocks?: number[];
+            specific?: {
+                conservative?: boolean;
+                data?: string;
+                from?: string;
+                to?: string;
+                txsize?: number;
+            };
+            feeLevels?: 'preloaded' | 'smart';
+        }
+    }
+
+    export interface FeeLevel {
+        label: 'high' | 'normal' | 'economy' | 'low' | 'custom';
+        feePerUnit: string;
+        blocks: number;
+        feeLimit?: string; // eth gas limit
+        feePerTx?: string; // fee for BlockchainEstimateFeeParams.request.specific
+    }
+
+    interface BlockchainEstimateFeeResponse {
+        blockTime: number;
+        minFee: number;
+        maxFee: number;
+        levels: FeeLevel[];
+    }
 
     namespace TrezorConnect {
         /**
          * Initializes TrezorConnect.
          */
-        function init(settings: Settings): void;
+        function init(settings: Settings): Promise<void>;
 
         /**
          * Retrieves BIP32 extended public derived by given BIP32 path.
@@ -859,9 +932,14 @@ declare module 'trezor-connect' {
          * returned in hexadecimal format. Change output is added automatically, if
          * needed.
          */
+        
+
         function composeTransaction(
             params: ComposeTransactionParams,
         ): Promise<ResponseMessage<Transaction>>;
+        function composeTransaction(
+            params: PrecomposeTransactionParams
+        ): Promise<ResponseMessage<PrecomposedTransaction[]>>;
 
         /**
          * Asks device to sign given inputs and outputs of pre-composed transaction.
@@ -927,6 +1005,8 @@ declare module 'trezor-connect' {
         function disableWebUSB(): void;
 
         function blockchainSubscribe(params: BlockchainSubscribeParams): Promise<ResponseMessage<BlockchainSubscribeResponse>>;
+
+        function blockchainEstimateFee(params: BlockchainEstimateFeeParams): Promise<ResponseMessage<BlockchainEstimateFeeResponse>>;
     }
 
     export default TrezorConnect;